fix: Code refactoring - Meeds-io/MIPs#121
- Services and REST Endpoints using Spring
- Js modules configuration bug fix
