+++ gamification-crowdin-services/src/main/java/io/meeds/gamification/crowdin/storage/CrowdinConsumerStorage.java
@@ -1,3 +1,20 @@
+/*
+ * This file is part of the Meeds project (https://meeds.io/).
+ *
+ * Copyright (C) 2023 Meeds Lab contact@meedslab.com
+ *
+ * This program is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 3 of the License, or (at your option) any later version.
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the GNU
+ * Lesser General Public License for more details.
+ * You should have received a copy of the GNU Lesser General Public License
+ * along with this program; if not, write to the Free Software Foundation,
+ * Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
+ */
 package io.meeds.gamification.crowdin.storage;
 
 import io.meeds.gamification.crowdin.exception.CrowdinConnectionException;
@@ -24,6 +41,7 @@ import org.exoplatform.services.log.ExoLogger;
 import org.exoplatform.services.log.Log;
 import org.json.JSONArray;
 import org.json.JSONObject;
+import org.springframework.stereotype.Component;
 
 import javax.ws.rs.core.MediaType;
 import java.io.IOException;
@@ -34,7 +52,7 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static io.meeds.gamification.crowdin.utils.Utils.*;
-
+@Component
 public class CrowdinConsumerStorage {
 
     private HttpClient client;
@@ -245,7 +263,7 @@ public class CrowdinConsumerStorage {
         }
     }
 
-    public String deleteWebhookHook(WebHook webHook) {
+    public String deleteWebhook(WebHook webHook) {
         URI uri = URI.create(CROWDIN_API_URL + PROJECTS + webHook.getProjectId() + WEBHOOKS + webHook.getWebhookId());
         try {
             return processDelete(uri, webHook.getToken());
