+++ gamification-crowdin-services/src/main/resources/conf/portal/configuration.xml
@@ -22,32 +22,41 @@
                xsi:schemaLocation="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd http://www.exoplatform.org/xml/ns/kernel_1_2.xsd"
                xmlns="http://www.exoplatform.org/xml/ns/kernel_1_2.xsd">
 
+
     <component>
-        <type>io.meeds.gamification.crowdin.rest.HooksManagementRest</type>
+        <key>CrowdinConnectorFeatureProperties</key>
+        <type>org.exoplatform.container.ExtendedPropertyConfigurator</type>
+        <init-params>
+           <properties-param>
+              <name>CrowdinConnectorFeatureProperties</name>
+              <description>Crowdin Connector Feature enablement flag</description>
+              <property name="exo.feature.CrowdinConnector.enabled" value="${exo.feature.CrowdinConnector.enabled:false}"/>
+           </properties-param>
+        </init-params>
     </component>
 
-  <component>
-    <key>CrowdinConnectorFeatureProperties</key>
-    <type>org.exoplatform.container.ExtendedPropertyConfigurator</type>
-    <init-params>
-       <properties-param>
-          <name>CrowdinConnectorFeatureProperties</name>
-          <description>Crowdin Connector Feature enablement flag</description>
-          <property name="exo.feature.CrowdinConnector.enabled" value="${exo.feature.CrowdinConnector.enabled:false}"/>
-       </properties-param>
-    </init-params>
-  </component>
+    <external-component-plugins>
+        <target-component>io.meeds.gamification.service.ConnectorService</target-component>
+        <component-plugin>
+            <name>crowdin</name>
+            <set-method>addPlugin</set-method>
+            <type>io.meeds.gamification.crowdin.plugin.CrowdinConnectorPlugin</type>
+        </component-plugin>
+    </external-component-plugins>
 
-  <external-component-plugins>
-    <target-component>io.meeds.gamification.service.ConnectorService</target-component>
-    <component-plugin>
-      <name>crowdin</name>
-      <set-method>addPlugin</set-method>
-      <type>io.meeds.gamification.crowdin.plugin.CrowdinConnectorPlugin</type>
-    </component-plugin>
-  </external-component-plugins>
-
-  <import>jar:/conf/portal/gamification-crowdin-connector-configuration.xml</import>
-  <import>jar:/conf/portal/gamification-crowdin-storage-configuration.xml</import>
-
+    <external-component-plugins>
+        <target-component>org.exoplatform.commons.api.persistence.DataInitializer</target-component>
+        <component-plugin>
+            <name>GamificationChangeLogsPlugin</name>
+            <set-method>addChangeLogsPlugin</set-method>
+            <type>org.exoplatform.commons.persistence.impl.ChangeLogsPlugin</type>
+            <init-params>
+                <values-param>
+                    <name>changelogs</name>
+                    <description>Change logs of Crowdin Connector</description>
+                    <value>db/changelog/crowdin-connector.db.changelog-1.0.0.xml</value>
+                </values-param>
+            </init-params>
+        </component-plugin>
+    </external-component-plugins>
 </configuration>
