+++ gamification-crowdin-webapp/src/main/resources/locale/portlet/CrowdinWebHookManagement_en.properties
@@ -40,7 +40,7 @@ crowdinConnector.admin.label.accessToken.instructions.stepOneLink=Crowdin settin
 #crowdinConnector.admin.label.accessToken.instructions.stepTwo=Be aware that your token can expire so ensure that the expiration date is set as expected
 crowdinConnector.admin.label.accessToken.instructions.stepThree=When creating the token, please make sure these scopes are selected
 
-crowdinConnector.admin.label.projectToWatch=You have set the remoteProjects to watch.
+crowdinConnector.admin.label.projectToWatch=You have set the projects to watch.
 crowdinConnector.admin.label.letEnableConnection=Now, let's enable the connection
 crowdinConnector.admin.label.enableConnection=Enable connection
 crowdinConnector.admin.label.letAllowConnection=Now, let's allow your contributors to connect their profile
