+++ gamification-crowdin-webapp/src/main/webapp/WEB-INF/gatein-resources.xml
@@ -23,5 +23,72 @@
                   xsi:schemaLocation="http://www.exoplatform.org/xml/ns/gatein_resources_1_4 http://www.exoplatform.org/xml/ns/gatein_resources_1_4"
                   xmlns="http://www.exoplatform.org/xml/ns/gatein_resources_1_4">
 
+
+    <module>
+        <name>gamificationUserConnectorsExtensionsCrowdin</name>
+        <load-group>gamification-user-connectors</load-group>
+        <script>
+            <path>/js/crowdinUserConnectorExtension.bundle.js</path>
+        </script>
+        <depends>
+            <module>vue</module>
+        </depends>
+        <depends>
+            <module>vuetify</module>
+        </depends>
+        <depends>
+            <module>eXoVueI18n</module>
+        </depends>
+        <depends>
+            <module>extensionRegistry</module>
+        </depends>
+        <depends>
+            <module>commonVueComponents</module>
+        </depends>
+    </module>
+
+    <module>
+        <name>gamificationAdminConnectorsExtensionsCrowdin</name>
+        <load-group>gamification-admin-connectors</load-group>
+        <script>
+            <path>/js/crowdinAdminConnectorExtension.bundle.js</path>
+        </script>
+        <depends>
+            <module>vue</module>
+        </depends>
+        <depends>
+            <module>vuetify</module>
+        </depends>
+        <depends>
+            <module>eXoVueI18n</module>
+        </depends>
+        <depends>
+            <module>extensionRegistry</module>
+        </depends>
+        <depends>
+            <module>commonVueComponents</module>
+        </depends>
+    </module>
+
+    <module>
+        <name>engagementCenterConnectorsCrowdinExtensions</name>
+        <load-group>engagement-center-connector-extensions</load-group>
+        <script>
+            <path>/js/connectorExtensions.bundle.js</path>
+        </script>
+        <depends>
+            <module>vue</module>
+        </depends>
+        <depends>
+            <module>vuetify</module>
+        </depends>
+        <depends>
+            <module>eXoVueI18n</module>
+        </depends>
+        <depends>
+            <module>extensionRegistry</module>
+        </depends>
+    </module>
+
 </gatein-resources>
